Change window splitting behaviour
When splitting a window, the new pane is now opened at the same working
directory as the existing pane. This should make it less of a hassle, as
we often want to be in the same place.

(Same behaviour can also be specified for opening new windows, but for
now I've left it as default.)